Moved some code to the correct object.
