+++ Source/ACE/Network/GameAction/Actions/GameActionChangeCombatMode.cs
@@ -10,7 +10,7 @@ namespace ACE.Network.GameAction.Actions
         public static void Handle(ClientMessage message, Session session)
         {
             var newCombatMode = message.Payload.ReadUInt32();
-            session.Player.SetCombatMode((CombatMode)newCombatMode, session.Player);
+            session.Player.SetCombatMode((CombatMode)newCombatMode);
         }
     }
 }
