+++ Source/ACE/Network/GameMessages/Messages/GameMessagePickupEvent.cs
@@ -6,11 +6,11 @@ namespace ACE.Network.GameMessages.Messages
 {
     public class GameMessagePickupEvent : GameMessage
     {
-        public GameMessagePickupEvent(Session session, WorldObject targetItem)
+        public GameMessagePickupEvent(ObjectGuid guid, WorldObject targetItem)
             : base(GameMessageOpcode.PickupEvent, GameMessageGroup.Group0A)
         {
             Writer.Write(targetItem.Guid.Full);
-            Writer.Write(session.Player.Guid.Full);
+            Writer.Write(guid.Full);
             Writer.Write(targetItem.Sequences.GetCurrentSequence(SequenceType.ObjectInstance));
             Writer.Write(targetItem.Sequences.GetCurrentSequence(SequenceType.ObjectPosition));
         }
